Move email files
