+++ backend/utils/emails/templates/forgotPassEmailTxt.js
