+++ backend/utils/emails/templates/verifyEmailTxt.js
